pwsearch/requirements.txt
colorama
gevent
grequests
loguru
pymediawiki
requests
rich